requirements.txt
uvicorn
fastapi
openai
python-dotenv
faiss-cpu
sentence-transformers
bs4
fake_useragent
lxml